+++ src/Console/PatternGenerator.php
@@ -16,7 +16,9 @@ class PatternGenerator extends Command
      *
      * @var string
      */
-    protected $signature = 'pattern:generate {name : Class (singular) for example User} {--M|model : Whether the generator should generate a model}';
+    protected $signature = 'pattern:generate 
+                        {name : Class (singular) for example User} 
+                        {--model : Whether the generator should generate a model}';
 
     /**
      * The console command description.
@@ -24,6 +26,11 @@ class PatternGenerator extends Command
      * @var string
      */
     protected $description = 'Generate Laravel Repository Pattern';
+
+    /**
+     * @var array
+     */
+    protected $generatedFiles = [];
 
     /**
      * Create a new command instance.
@@ -41,33 +48,143 @@ class PatternGenerator extends Command
     public function handle()
     {
         $name = $this->argument('name');
-        $generateModel = $this->argument('model');
+        $generateModel = $this->option('model');
 
         $this->info('Generating...');
+        $this->copyStaticFiles();
 
-        if($generateModel) {
+        if ($generateModel) {
             $this->model($name);
+            $this->info('Model is ready!');
         }
 
         $this->controller($name);
+        $this->info('Controller is ready!');
+
         $this->apiController('v1', $name);
+        $this->info('API Controller is ready!');
+
         $this->request($name);
         $this->updateRequest($name);
+        $this->info('Controller requests are ready!');
+
         $this->eloquent($name);
         $this->interface($name);
+        $this->info('Repository layer is ready!');
+
         $this->service($name);
-        $this->viewModel($name);
-        $this->transformer($name);
+        $this->info('Controller service is ready!');
+
+        $this->provider($name);
+        $this->info('Provider is ready!');
+
+        $this->resource($name);
+        $this->info('Resource is ready!');
+
+        $this->filter($name);
+        $this->info('Filter is ready!');
+
         $this->views($name);
+        $this->info('Base views are ready!');
 
-        File::append(base_path('routes/api.php'), 'Route::resource(\'' . Str::plural(strtolower($name)) . "', {$name}ApiController::class)->only(['store', 'update', 'destroy']);\n\r");
-        File::append(base_path('routes/web.php'), 'Route::resource(\'' . Str::plural(strtolower($name)) . "', {$name}Controller::class)->only(['index', 'create', 'edit']);\n\r");
+        $this->routes($name);
+        $this->info('Routes are ready!');
 
-        $this->info('Generate finished!');
+        $this->newLine();
+        $this->info('Generated files:');
+        $this->table(
+            ['Name', 'Location'],
+            $this->generatedFiles
+        );
+
+        $this->newLine();
+        $this->info('Please add this line to config/app.php Application Service Providers section:');
+        $this->info("App\Domain\/".$name."\Providers\/".$name."ServiceProvider::class,");
+
+        $this->newLine();
+        $this->info('Please add this line to config/role-middlewares.php controllers section:');
+        $lowerName = strtolower($name);
+        $this->info("'{$lowerName}' => [],");
+
     }
 
     /**
-     * @param string $name
+     * Create static directories and copy static files
+     */
+    protected function copyStaticFiles()
+    {
+        //Create directories
+        if (!file_exists($path = app_path("Http/Controllers"))) {
+            mkdir($path, 0777, true);
+        }
+
+        if (!file_exists($path = app_path("Domain"))) {
+            mkdir($path, 0777, true);
+        }
+
+        //Copy static files
+        $baseController = $this->getStub('BaseTransactionController');
+        if (!file_exists($path = app_path("Http/Controllers/BaseTransactionController.php"))) {
+            file_put_contents($path, $baseController);
+            $this->generatedFiles[] = [
+                'name' => 'BaseTransactionController.php',
+                'location' => $path
+            ];
+        }
+        $this->info('Base Controller is ready!');
+
+        $abstractEloquentRepository = $this->getStub('AbstractEloquentRepository');
+        if (!file_exists($path = app_path("Domain/AbstractEloquentRepository.php"))) {
+            file_put_contents($path, $abstractEloquentRepository);
+            $this->generatedFiles[] = [
+                'name' => 'AbstractEloquentRepository.php',
+                'location' => $path
+            ];
+        }
+        $this->info('AbstractEloquentRepository is ready!');
+
+        $abstractQueryFilter = $this->getStub('AbstractQueryFilter');
+        if (!file_exists($path = app_path("Domain/AbstractQueryFilter.php"))) {
+            file_put_contents($path, $abstractQueryFilter);
+            $this->generatedFiles[] = [
+                'name' => 'AbstractQueryFilter.php',
+                'location' => $path
+            ];
+        }
+        $this->info('AbstractQueryFilter is ready!');
+
+        $eloquentRepositoryInterface = $this->getStub('EloquentRepositoryInterface');
+        if (!file_exists($path = app_path("Domain/EloquentRepositoryInterface.php"))) {
+            file_put_contents($path, $eloquentRepositoryInterface);
+            $this->generatedFiles[] = [
+                'name' => 'EloquentRepositoryInterface.php',
+                'location' => $path
+            ];
+        }
+        $this->info('EloquentRepositoryInterface is ready!');
+
+        $repositoryInterface = $this->getStub('RepositoryInterface');
+        if (!file_exists($path = app_path("Domain/RepositoryInterface.php"))) {
+            file_put_contents($path, $repositoryInterface);
+            $this->generatedFiles[] = [
+                'name' => 'RepositoryInterface.php',
+                'location' => $path
+            ];
+        }
+        $this->info('RepositoryInterface is ready!');
+    }
+
+    /**
+     * @param  string  $name
+     * @return false|string
+     */
+    protected function getStub(string $name)
+    {
+        return file_get_contents(__DIR__."/stubs/$name.stub");
+    }
+
+    /**
+     * @param  string  $name
      */
     protected function model(string $name)
     {
@@ -77,20 +194,15 @@ class PatternGenerator extends Command
             $this->getStub('Model')
         );
 
-        file_put_contents(app_path("Models/{$name}.php"), $modelTemplate);
+        file_put_contents($path = app_path("Models/{$name}.php"), $modelTemplate);
+        $this->generatedFiles[] = [
+            'name' => "{$name}.php",
+            'location' => $path
+        ];
     }
 
     /**
-     * @param string $name
-     * @return false|string
-     */
-    protected function getStub(string $name)
-    {
-        return file_get_contents(__DIR__ . "/stubs/$name.stub");
-    }
-
-    /**
-     * @param string $name
+     * @param  string  $name
      */
     protected function controller(string $name)
     {
@@ -112,12 +224,16 @@ class PatternGenerator extends Command
             mkdir($path, 0777, true);
         }
 
-        file_put_contents(app_path("Http/Controllers/Web/{$name}Controller.php"), $controllerTemplate);
+        file_put_contents($path = app_path("Http/Controllers/Web/{$name}Controller.php"), $controllerTemplate);
+        $this->generatedFiles[] = [
+            'name' => "{$name}Controller.php",
+            'location' => $path
+        ];
     }
 
     /**
-     * @param string $version
-     * @param string $name
+     * @param  string  $version
+     * @param  string  $name
      */
     protected function apiController(string $version, string $name)
     {
@@ -139,11 +255,16 @@ class PatternGenerator extends Command
             mkdir($path, 0777, true);
         }
 
-        file_put_contents(app_path("Http/Controllers/Api/{$version}/{$name}ApiController.php"), $apiControllerTemplate);
+        file_put_contents($path = app_path("Http/Controllers/Api/{$version}/{$name}ApiController.php"),
+            $apiControllerTemplate);
+        $this->generatedFiles[] = [
+            'name' => "{$name}ApiController.php",
+            'location' => $path
+        ];
     }
 
     /**
-     * @param string $name
+     * @param  string  $name
      */
     protected function request(string $name)
     {
@@ -157,11 +278,15 @@ class PatternGenerator extends Command
             mkdir($path, 0777, true);
         }
 
-        file_put_contents(app_path("Http/Requests/{$name}Request.php"), $requestTemplate);
+        file_put_contents($path = app_path("Http/Requests/{$name}Request.php"), $requestTemplate);
+        $this->generatedFiles[] = [
+            'name' => "{$name}Request.php",
+            'location' => $path
+        ];
     }
 
     /**
-     * @param string $name
+     * @param  string  $name
      */
     protected function updateRequest(string $name)
     {
@@ -175,11 +300,15 @@ class PatternGenerator extends Command
             mkdir($path, 0777, true);
         }
 
-        file_put_contents(app_path("Http/Requests/{$name}UpdateRequest.php"), $updateRequestTemplate);
+        file_put_contents($path = app_path("Http/Requests/{$name}UpdateRequest.php"), $updateRequestTemplate);
+        $this->generatedFiles[] = [
+            'name' => "{$name}UpdateRequest.php",
+            'location' => $path
+        ];
     }
 
     /**
-     * @param string $name
+     * @param  string  $name
      */
     protected function eloquent(string $name)
     {
@@ -201,11 +330,16 @@ class PatternGenerator extends Command
             mkdir($path, 0777, true);
         }
 
-        file_put_contents(app_path("Domain/{$name}/Repositories/Eloquent{$name}Repository.php"), $eloquentTemplate);
+        file_put_contents($path = app_path("Domain/{$name}/Repositories/Eloquent{$name}Repository.php"),
+            $eloquentTemplate);
+        $this->generatedFiles[] = [
+            'name' => "Eloquent{$name}Repository.php",
+            'location' => $path
+        ];
     }
 
     /**
-     * @param string $name
+     * @param  string  $name
      */
     protected function interface(string $name)
     {
@@ -227,11 +361,16 @@ class PatternGenerator extends Command
             mkdir($path, 0777, true);
         }
 
-        file_put_contents(app_path("Domain/{$name}/Repositories/Interfaces/{$name}RepositoryInterface.php"), $interfaceTemplate);
+        file_put_contents($path = app_path("Domain/{$name}/Repositories/Interfaces/{$name}RepositoryInterface.php"),
+            $interfaceTemplate);
+        $this->generatedFiles[] = [
+            'name' => "{$name}RepositoryInterface.php",
+            'location' => $path
+        ];
     }
 
     /**
-     * @param string $name
+     * @param  string  $name
      */
     protected function service(string $name)
     {
@@ -253,63 +392,87 @@ class PatternGenerator extends Command
             mkdir($path, 0777, true);
         }
 
-        file_put_contents(app_path("Domain/{$name}/Services/{$name}Service.php"), $serviceTemplate);
+        file_put_contents($path = app_path("Domain/{$name}/Services/{$name}Service.php"), $serviceTemplate);
+        $this->generatedFiles[] = [
+            'name' => "{$name}Service.php",
+            'location' => $path
+        ];
     }
 
     /**
-     * @param string $name
+     * @param  string  $name
      */
-    protected function viewModel(string $name)
+    protected function provider(string $name)
     {
-        $viewModelTemplate = str_replace(
+        $providerTemplate = str_replace(
             [
                 '{{modelName}}',
-                '{{modelNamePluralLowerCase}}',
                 '{{modelNameSingularLowerCase}}',
             ],
             [
                 $name,
-                strtolower(Str::plural($name)),
                 strtolower($name),
             ],
-            $this->getStub('ViewModel')
+            $this->getStub('Provider')
         );
 
-        if (!file_exists($path = app_path("Domain/{$name}/ViewModel"))) {
+        if (!file_exists($path = app_path("Domain/{$name}/Providers"))) {
             mkdir($path, 0777, true);
         }
 
-        file_put_contents(app_path("Domain/{$name}/ViewModel/{$name}.php"), $viewModelTemplate);
+        file_put_contents($path = app_path("Domain/{$name}/Providers/{$name}ServiceProvider.php"), $providerTemplate);
+        $this->generatedFiles[] = [
+            'name' => "{$name}ServiceProvider.php",
+            'location' => $path
+        ];
     }
 
     /**
-     * @param string $name
+     * @param  string  $name
      */
-    protected function transformer(string $name)
+    protected function resource(string $name)
     {
-        $transformerTemplate = str_replace(
-            [
-                '{{modelName}}',
-                '{{modelNamePluralLowerCase}}',
-                '{{modelNameSingularLowerCase}}',
-            ],
-            [
-                $name,
-                strtolower(Str::plural($name)),
-                strtolower($name),
-            ],
-            $this->getStub('Transformer')
+        $resourceTemplate = str_replace(
+            ['{{modelName}}'],
+            [$name],
+            $this->getStub('Resource')
         );
 
-        if (!file_exists($path = app_path("Domain/{$name}/ViewModel/Transformers"))) {
+        if (!file_exists($path = app_path("Domain/{$name}/Resources"))) {
             mkdir($path, 0777, true);
         }
 
-        file_put_contents(app_path("Domain/{$name}/ViewModel/Transformers/{$name}Transformer.php"), $transformerTemplate);
+        file_put_contents($path = app_path("Domain/{$name}/Resources/{$name}Resource.php"), $resourceTemplate);
+        $this->generatedFiles[] = [
+            'name' => "{$name}Resource.php",
+            'location' => $path
+        ];
     }
 
     /**
-     * @param string $name
+     * @param  string  $name
+     */
+    protected function filter(string $name)
+    {
+        $filterTemplate = str_replace(
+            ['{{modelName}}'],
+            [$name],
+            $this->getStub('Filter')
+        );
+
+        if (!file_exists($path = app_path("Domain/{$name}/Filters"))) {
+            mkdir($path, 0777, true);
+        }
+
+        file_put_contents($path = app_path("Domain/{$name}/Filters/{$name}Filter.php"), $filterTemplate);
+        $this->generatedFiles[] = [
+            'name' => "{$name}Filter.php",
+            'location' => $path
+        ];
+    }
+
+    /**
+     * @param  string  $name
      */
     protected function views(string $name)
     {
@@ -335,8 +498,48 @@ class PatternGenerator extends Command
             mkdir($path, 0777, true);
         }
 
-        file_put_contents(resource_path("views/{$lower}/index.blade.php"), $viewIndexTemplate);
-        file_put_contents(resource_path("views/{$lower}/create.blade.php"), $viewEmptyTemplate);
-        file_put_contents(resource_path("views/{$lower}/edit.blade.php"), $viewEmptyTemplate);
+        file_put_contents($path = resource_path("views/{$lower}/index.blade.php"), $viewIndexTemplate);
+        $this->generatedFiles[] = [
+            'name' => "views/{$lower}/index.blade.php",
+            'location' => $path
+        ];
+
+        file_put_contents($path = resource_path("views/{$lower}/create.blade.php"), $viewEmptyTemplate);
+        $this->generatedFiles[] = [
+            'name' => "views/{$lower}/create.blade.php",
+            'location' => $path
+        ];
+
+        file_put_contents($path = resource_path("views/{$lower}/edit.blade.php"), $viewEmptyTemplate);
+        $this->generatedFiles[] = [
+            'name' => "views/{$lower}/edit.blade.php",
+            'location' => $path
+        ];
+    }
+
+    /**
+     * @param  string  $name
+     */
+    protected function routes(string $name)
+    {
+        File::append(base_path('routes/web.php'),
+            'Route::resource(\''.Str::plural(strtolower($name))."', {$name}Controller::class)->only(['index', 'create', 'edit']);\n\r");
+        $this->generatedFiles[] = [
+            'name' => "web.php",
+            'location' => base_path('routes/web.php')
+        ];
+
+        File::append(base_path('routes/api.php'), 'Route::resource(\''.Str::plural(strtolower($name))."', {$name}ApiController::class, [
+            'names' => [
+                'index' => 'api.".Str::plural(strtolower($name)).".index',
+                'store' => 'api.".Str::plural(strtolower($name)).".store',
+                'update' => 'api.".Str::plural(strtolower($name)).".update',
+                'destroy' => 'api.".Str::plural(strtolower($name)).".destroy',
+            ],
+        ])->only(['index', 'store', 'update', 'destroy']);\n\r");
+        $this->generatedFiles[] = [
+            'name' => "api.php",
+            'location' => base_path('routes/api.php')
+        ];
     }
 }